tunned some variables for better training, and made the actor validate the probs before selecting one
stopped checking for action validty in the env. and tunned reward of distance with 10km instead of 5, dicreased learning rate of actor and critic, which helps a lot, and changed maximum agent_num_updates to be 10 instead of 5. added cluster to the
testing case, and finally, tunned gamma to be .95 instead of .99